Only show the tooltip when we are no top of the sample box in sample graph
Previously we were always returning the closest sample to the mouse
hover/click location. Which was okay when we were clicking to select the
boxes before. But now we are also showing tooltips on hover which can be
misleading if they were shown when we are not hovering over them. So to
make this experience less confusing, this commit changes the behavior to
only select or hover a specific sample when the mouse is actually on top
of the sample box that we draw.