Move the profile startTime calculation to somewhere common
Previously we were only doing this for CPUProfile profiles. But chrome
adds this timestamp to all of its profile files including the Tracing
files. See their source code here:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/devtools-frontend/src/front_end/panels/timeline/TimelinePanel.ts;l=1242-1251;drc=0cc387000bc7c9369fc34c2f2390d3cbc67f65ab
So that's why this commit makes sure that we check the profile url for
every type of chrome profiles, and not only for CPU profiles.