Bring back the profilingStartTime and shift the startTime depending on that
We need the profilingStartTime for calculating the time range. That's
why we can't simply put zero. But also we need to make sure that
profiler frontend reports the correct time. Profiler calculates the
recording start timestamp as `startTime + profilingStartTime`. So to be
able to tell the user about the correct timestamp we read from the
profile file name, we have to calculate the `startTime` like this:
`timestamp from name - profilingStartTime`. This commit changes this to
make sure that we always show the correct timestamp.