Fix version 41 upgrader.
This upgrader expects thread.stringTable to be a StringTable.
But upgraders run on the serialized version of the profile,
so threads only have a stringArray, and no stringTable. So this
code could never have worked. This must mean that the code path
isn't taken very often, i.e. that it's unlikely to encounter a
resource without a name.

Even so, it's better to fix it so that future readers don't get confused.