Remove special handling of tracing markers when looking up marker schemas.
This adds an upgrader so that old profiles still get their CC markers
displayed in the memory track and non-CC tracing markers displayed in
the timeline-overview area.

Newer profiles don't need this special case because their CC markers have
a CC schema which correctly puts them in the memory track, and because
they include a 'tracing' schema which puts non-CC tracing markers into
the timeline-overview area.