Make sample indexes compatible between the unfiltered and (preview) filtered call tree summary strategy samples when using an allocation strategy.
Fixes #5331.

Let's say we allocate 100 objects, and then free the first 50 of them.
If you look at the entire thread, the `retained-native-allocations` view
will show the allocation samples for the second 50 objects (the ones
that weren't freed).
If you look at just the part of the thread where you allocated, you
will see allocation samples for all 100 objects (because none were
freed in that part of the thread).

How do you get the stack category for each of the 100 allocation samples,
especially if you've applied transforms or a JS-only filter?

You need to look at the stack of the raw nativeAllocations table.

You cannot look at the `retained-native-allocations` of the entire
thread, because that table doesn't have the stacks for the first
50 objects - it removes those samples because it knows that those
objects have been deallocated.

This commit makes it so that the unfilteredCtssSamples are just
the raw allocation tables, when an allocation-related call tree
summary strategy is selected.

And it makes it so that `extractSamplesLikeTable` always returns
a table that has compatible indexes to the `extractUnfilteredCtssSamples`,
just with some samples having a null stack.

Now the unfilteredCtssSamples can be used to look up the stack, and
therefore the original category of a sample, based on the sample
index in the filtered (or preview-filtered) CTSS samples table. We
should not create a call tree just based on the unfilteredCtssSamples
because these samples have not been filtered according to the strategy.