Add a compacting step at the end of profile sanitization.
In the future we're planning to share more tables across threads,
starting with the string table.

If a user wants to upload only visible threads, there's the risk of
keeping data from other threads in the profile if more of that data
is shared across threads.

So this commit adds a "compacting garbage collection" step before
uploading, so that only data which is referenced within the sanitized
profile is kept in the profile.