Compute the "scroll to hot spot" when making the bottom box info.
This means we can compute the right line number for the "hot spot" for the
call node that was passed to getBottomBoxInfoForCallNode, and not just for
the selected call node.

For example, getBottomBoxInfoForCallNode can be called with the right-clicked
call node which can be different from the selected call node, and we were
scrolling to the wrong spot in that case before this patch.

Fixes #5758.