Hardcode the marker schema in this test which creates a profile of version 33.
When the test was initially added, the markerSchema was missing from the profile.
That was invalid because version 33 is the version which started requiring that
the markerSchema field is present.

In #4352 we added the missing markerSchema because the upgrader for
version 44 caught the fact that it was missing, because it (rightfully) assumes
that it's present.

However, since this is a profile of a fixed version, we shouldn't attempt to
share code with other tests - anything that's shared will usually follow the
most recent format version, whereas this test must make sure the profile it
creates complies with the hardcoded fixed version.