Use RawThread in getThreadProcessDetails.
This means that we don't need to compute the derived thread for non-visible threads.
We call getThreadProcessDetails on non-visible threads when deriving the props for
TimelineLocalTrack components.

Furthermore, this speeds up calls to getThreadProcessDetails after redux state changes,
because getRawThread doesn't have any dependent selectors that need to be revalidated.