Change the test expectations to any co2 values
Previously we had a specific value for the expected co2 number, which
was problematic because whenever we had to upgrade the co2 package, we
had to update these values too. It's a small annoyance, but it can be
still annoying, especially you had to stop what you were doing and just
update the values of the tests.

So instead now we expect any values for these tests expectations. That
way we can upgrade the co2 package without problem!