Add a more granular test for bandwidth tooltips
Similar reason to the previous commit, but we had to change this
snapshot test into a more granular test that checks individual fields
because it's not very easy to get the html and then replace all the text
contents, similar to `getBy*`. This is a lot safer to do instead.

We could simply get the html content with `.innerHtml` but it doesn't
pretty print the DOM tree, which makes it very difficult to understand
the changes. There is a `prettyDOM` method in testing-library, but it's
added for debugging purposes and not really for testing.