Simplify the active track left border css and make it right clickable
Previously it was doing a lot more `calc`s because it was happening in
the parent div. This commit changes the box-shadow into an inset
box-shadow so we don't have to change the width all the time. But we
still need to change the width once to remove the left border for the
local tracks.