Don't warn when we can't find the navigation start marker.
Previously the condition `if (tabThreadStringTable.hasString('Navigation::Start'))`
was a cheap way to answer the question `do we have such a marker in this thread?`,
so the error made sense because this shouldn't have happened. But now this will
happen in all threads as soon as one thread has one of these.