Patch jsdom to to make window.location configurable
Jest version 30 page mentions this as a known issue:
https://jestjs.io/blog/2025/06/04/jest-30#known-issues

It looks like the new version of jsdom made window.location
non-configurable. This is not great because we mock it in our tests
since jsdom/jest doesn't provide everything we want.

Unfortunately jsdom is not very positive on fixing this. See:
https://github.com/jsdom/jsdom/issues/3492