Sort the threads that have different name too and fall back to original sorting behavior
This created some test changes how we find the selected thread in comparison
profiles especially. But after thinking about it, I decided that the new
behavior is actually better for them. Previously we were always getting the
first track, but now we are getting the crowded one, which is usually the
diffing track.