Remove unintended gaps on small screens.
On small screens, the top bar was overflowing to the right, making the entire
document scrollable, which lead to unintended dark blue areas appearing.

As a quick fix, make the bar itself scrollable.

Also put a minimum size on the profile name button and make sure its edit icon
isn't squished.

And prevent scrolling the contents of .Details, for example from a single overflowing
pixel.