Move L10nState from Redux to React component state
Most importantly, this means we'll no longer store a ReactLocalization
object in the Redux store. This makes it easier to use our Redux
state and selectors from a library that runs outside of a node context.

We weren't really using the Redux facilities for anything other than
value storage. The localization, requestedLocales and direction properties
were only needed locally within the AppLocalizationProvider component.
The primaryLocale and requestL10n properties are also referenced from
the LanguageSwitcher component; we're now using React context and a
useL10n hook to keep that working.
And for the togglePseudoStrategy console command we use a hacky exported
function that accesses a global.

Overall I think this is simpler; it certainly removes a whole bunch of code.