Make sure that the test-debug command runs the tests properly
We should set these environment variables to make sure that all the
tests behave the same way they would with `yarn test`.

Also, we added this check so we make sure people run the tests through
`yarn test`:
https://github.com/firefox-devtools/profiler/blob/082631b9ab95733b5e1e36b5edfb73cbcedcd866/src/test/setup.js#L19-L21
But it was preventing us to run `yarn test-debug` all together too.