Add Node compatibility to gz.js.
Node has a built-in zlib library we can just use.

This also changes the webpack config to tell it that it we don't need
this library when building for the browser. Without this change, the
build fails because Webpack wants to know explicitly what, if anything,
it should shim the library with.