Change Worker mocking setup.
Make it so that gz.ts can just use `new Worker` as if it was running in
the browser. During tests, we mock window.Worker with our NodeWorker
implementation.

Replace the manual name-to-path mapping with the use of Webpack's file-loader.

Using Webpack's file loader is a more natural way to define which files
we depend on. This means that, if our this code is used as part of an
npm library, the Webpack build output will contain the worker files, and
the npm library should find it - as long as it's running in the browser.
(I haven't verified this though; I'm not sure if extra resource files from
npm dependencies are carried through into the final artifact properly.)

As a follow-up, we can add a Node compatibility path to gz.js, for when this
code is used as part of Node library, i.e. actually running within Node.
In our tests we'd probably still want to use our Worker mock so that we test
the code we ship to the browser.