Remove the unneeded @babel/no-invalid-this eslint disable rule
It looks like this line was a false positive for @babel/no-invalid-this
initially because here `this` comes from a parameter and it's not a
global `this`. After we upgraded the eslint version, this is correctly
caught and not reported as error.

I double checked this rule, by specifically making this mistake in the
codebase, and it looks like we catch it properly elsewhere, so it's not
a case where we fail to detect this rule.

This is the explanation of that rule:
https://eslint.org/docs/latest/rules/no-invalid-this

I explicitly tried the first incorrect example here and `yarn lint-js`
catches it properly.