Move the simpleperf magic value check.
esbuild only supports code splitting in the configuration where we ship
actual ES modules to the browser. In that configuration, if a module is
used by both the main chunk and an async chunk, that module becomes a
separate chunk. That's rather sillyl for a module that's just a single
constant. So we just copy the constant and remove the import.