Restore correct stack substitution code.
We need to replace cause.stack even if convertStack returned null,
otherwise we have bad stack indexes sticking around.

The incorrect `if (stack)` was added during the TypeScript migration
in commit ed02f3f5820313da1e9f55db94054500b0612b89.

This restores the logic as it was before the migration.