Add sources table to the gecko and processed profile formats
This patch adds the sources table but it doesn't use that yet. The
following commits will be changing the funcTable and then start using
it.

Note that this gathers all sources, including the JS and native sources.
This way, we have a single source of truth for them. But this will
happen in the following patches.

This patch:
- Adds a `geckoProfile.sources` table to the gecko profile format.
- Handles the processing of gecko profile -> processed profile format.