Extract the GlobalDataCollector class to its own file
This is needed for the following patches to make sure that we don't have
a circular dependency. Previously this class was only used for our own
profile processing from the gecko profile data. But in the next commits
I'm changing the profile importers to use this as well for easy profile
shared data collection.