Add type annotations to all React component state members.
I'm not really sure why, but without this, TypeScript thinks that the state
can only have the values that it sees in the initialization value. It also
doesn't complain if setState is called with values that don't match the type
it inferred for the state member.

It needs to know that this.state can have other values than the initial value
if we want to avoid false positives from '@typescript-eslint/no-unnecessary-condition'.