Remove the unused codecov dependency
This dependency was only used inside the CircleCI task for uploading the
code coverage results with `yarn codecov`. Since we use the official
codecov action in GitHub Actions now, we don't really need it anymore.

Note that running `yarn codecov` locally is not possible as it will
error out since we are not in a CI environment.