Redirect unpublished url loads to the homepage similar to from-file (#5712)
Initially `unpublished` was an alias of `from-addon` and then it was
made a separate data source. But reloading the url or loading from
scratch was hanging in the profile loading animation indefinitely and it
wasn't trying to load any profile either. It makes sense to make it
behave more like `from-file` in that sense and just redirect to the home
page.

Issue #5709 has an STR for this bug.