Move symbol table caching and demangling into RegularSymbolProvider.
This removes the demangling dependency from symbol-store.ts.
It also lets us remove the unused "database" from symbolicator-cli.

The cache only exists so that we don't give the browser unnecessary
work to recompute the symbol table. So we only need to access the cache
if we're about to ask the browser for a full symbol table.